All parsing is well.
With the modifications made to the Noweb 2.13 sources, removing all the
instances of the deprecated "@ %def IDENTIFIER" syntax, peg-noweb can
successfully parse all Noweb sources with the current grammar and can also
deparse such to reproduce the tool syntax.

Save for any program features not utilized, such as indexing and
cross-referencing and autodefinitions in the pipeline which produces the
document, all features have been tested.